test: :test_tube: reorganize unit tests for better structure
