+++ tests/Unit/DTO/Auth/RegisterUserDtoTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Unit;
+namespace Tests\Unit\DTO\Auth;
 
 use App\DTO\Auth\RegisterUserDTO;
 use PHPUnit\Framework\TestCase;
